Move git stuff to separate block
